--- src/core/embedders/CustomFontNonSubsetEmbedder.ts
@@ -1,41 +1,0 @@
-import { create as createFont } from '@denkiyagi/fontkit';
-import type { TTFFont } from '@denkiyagi/fontkit';
-
-import { CustomFontEmbedder } from 'src/core/embedders/CustomFontEmbedder';
-import type { EmbedFontAdvancedOptions } from 'src/api';
-
-export class CustomFontNonSubsetEmbedder extends CustomFontEmbedder {
-  static for(
-    fontData: Uint8Array,
-    customFontName?: string,
-    vertical?: boolean,
-    advanced?: EmbedFontAdvancedOptions,
-  ) {
-    const font = createFont(fontData);
-    if (font.type !== 'TTF') throw new Error(`Invalid font type: ${font.type}`);
-    return new CustomFontNonSubsetEmbedder(
-      font,
-      fontData,
-      customFontName,
-      vertical,
-      advanced,
-    );
-  }
-
-  private readonly fontData: Uint8Array;
-
-  private constructor(
-    font: TTFFont,
-    fontData: Uint8Array,
-    customFontName?: string,
-    vertical?: boolean,
-    advanced?: EmbedFontAdvancedOptions,
-  ) {
-    super(font, customFontName, vertical, advanced);
-    this.fontData = fontData;
-  }
-
-  protected serializeFont(): Uint8Array {
-    return this.fontData;
-  }
-}
