+++ src/core/embedders/CustomFontSubsetEmbedder.ts
@@ -1,7 +1,7 @@
 import { create as createFont, LayoutAdvancedParams } from '@denkiyagi/fontkit';
 import type { TTFFont, Glyph, Subset } from '@denkiyagi/fontkit';
 
-import { CustomFontEmbedder } from 'src/core/embedders/CustomFontEmbedder';
+import { AbstractCustomFontEmbedder } from 'src/core/embedders/AbstractCustomFontEmbedder';
 import { PDFHexString } from 'src/core/objects/PDFHexString';
 import { Cache, toHexStringOfMinLength } from 'src/utils';
 import type { EmbedFontAdvancedOptions } from 'src/api';
@@ -12,7 +12,7 @@ import type { SingleLineTextOrGlyphs } from 'src/types/text';
  * this class borrows from:
  *   https://github.com/devongovett/pdfkit/blob/e71edab0dd4657b5a767804ba86c94c58d01fbca/lib/image/jpeg.coffee
  */
-export class CustomFontSubsetEmbedder extends CustomFontEmbedder {
+export class CustomFontSubsetEmbedder extends AbstractCustomFontEmbedder {
   static for(
     fontData: Uint8Array,
     customFontName?: string,
