+++ tests/core/embedders/CustomFontEmbedder.spec.ts
@@ -1,7 +1,7 @@
 import fs from 'fs';
 
 import {
-  CustomFontNonSubsetEmbedder,
+  CustomFontEmbedder,
   PDFContext,
   PDFDict,
   PDFHexString,
@@ -10,14 +10,14 @@ import {
 
 const ubuntuFont = fs.readFileSync('./assets/fonts/ubuntu/Ubuntu-R.ttf');
 
-describe(`CustomFontNonSubsetEmbedder`, () => {
-  it(`can be constructed with CustomFontNonSubsetEmbedder.for(...)`, async () => {
-    const embedder = await CustomFontNonSubsetEmbedder.for(ubuntuFont);
-    expect(embedder).toBeInstanceOf(CustomFontNonSubsetEmbedder);
+describe(`CustomFontEmbedder`, () => {
+  it(`can be constructed with CustomFontEmbedder.for(...)`, async () => {
+    const embedder = await CustomFontEmbedder.for(ubuntuFont);
+    expect(embedder).toBeInstanceOf(CustomFontEmbedder);
   });
 
   it(`exposes the font's name`, async () => {
-    const embedder = await CustomFontNonSubsetEmbedder.for(
+    const embedder = await CustomFontEmbedder.for(
       new Uint8Array(ubuntuFont),
     );
     expect(embedder.fontName).toBe('Ubuntu');
@@ -25,7 +25,7 @@ describe(`CustomFontNonSubsetEmbedder`, () => {
 
   it(`can set a custom font name`, async () => {
     const customName = 'abc123';
-    const embedder = await CustomFontNonSubsetEmbedder.for(
+    const embedder = await CustomFontEmbedder.for(
       new Uint8Array(ubuntuFont),
       customName,
     );
@@ -34,7 +34,7 @@ describe(`CustomFontNonSubsetEmbedder`, () => {
 
   it(`can embed font dictionaries into PDFContexts without a predefined ref`, async () => {
     const context = PDFContext.create();
-    const embedder = await CustomFontNonSubsetEmbedder.for(
+    const embedder = await CustomFontEmbedder.for(
       new Uint8Array(ubuntuFont),
     );
 
@@ -47,7 +47,7 @@ describe(`CustomFontNonSubsetEmbedder`, () => {
   it(`can embed font dictionaries into PDFContexts with a predefined ref`, async () => {
     const context = PDFContext.create();
     const predefinedRef = PDFRef.of(9999);
-    const embedder = await CustomFontNonSubsetEmbedder.for(
+    const embedder = await CustomFontEmbedder.for(
       new Uint8Array(ubuntuFont),
     );
 
@@ -62,7 +62,7 @@ describe(`CustomFontNonSubsetEmbedder`, () => {
     const text = 'Stuff and thingz!';
     const hexCodes =
       '00360057005801AA000300440051004700030057004B004C0051004A005D0004';
-    const embedder = await CustomFontNonSubsetEmbedder.for(ubuntuFont);
+    const embedder = await CustomFontEmbedder.for(ubuntuFont);
 
     expect(embedder.encodeText(text)).toBeInstanceOf(PDFHexString);
     expect(String(embedder.encodeText(text))).toBe(
@@ -72,19 +72,19 @@ describe(`CustomFontNonSubsetEmbedder`, () => {
 
   it(`can measure the width of text strings at the given font size`, async () => {
     const text = 'Stuff and thingz!';
-    const embedder = await CustomFontNonSubsetEmbedder.for(ubuntuFont);
+    const embedder = await CustomFontEmbedder.for(ubuntuFont);
     expect(embedder.widthOfTextAtSize(text, 12)).toBe(90.672);
     expect(embedder.widthOfTextAtSize(text, 24)).toBe(181.344);
   });
 
   it(`can measure the height of the font at the given size`, async () => {
-    const embedder = await CustomFontNonSubsetEmbedder.for(ubuntuFont);
+    const embedder = await CustomFontEmbedder.for(ubuntuFont);
     expect(embedder.heightOfFontAtSize(12)).toBeCloseTo(13.452);
     expect(embedder.heightOfFontAtSize(24)).toBeCloseTo(26.904);
   });
 
   it(`can measure the size of the font at a given height`, async () => {
-    const embedder = await CustomFontNonSubsetEmbedder.for(ubuntuFont);
+    const embedder = await CustomFontEmbedder.for(ubuntuFont);
     expect(embedder.sizeOfFontAtHeight(12)).toBeCloseTo(10.705);
     expect(embedder.sizeOfFontAtHeight(24)).toBeCloseTo(21.409);
   });
